Reskin about section in settings
Moved credits to its own section
